Remove Ginkgo
