--- e2e/go.mod
@@ -1,15 +1,0 @@
-module github.com/cerbos/cerbos/e2e
-
-go 1.17
-
-require (
-	github.com/onsi/ginkgo/v2 v2.0.0-rc3
-	github.com/onsi/gomega v1.17.0
-)
-
-require (
-	golang.org/x/net v0.0.0-20210428140749-89ef3d95e781 // indirect
-	golang.org/x/sys v0.0.0-20210423082822-04245dca01da // indirect
-	golang.org/x/text v0.3.6 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
