+++ e2e/run.sh
@@ -7,7 +7,7 @@ CLUSTER=${CLUSTER:-"cerbos-e2e"}
 CLEANUP=${CLEANUP:-"true"}
 
 check_prerequisites() {
-    for EXE in ginkgo helm helmfile kind kubectl telepresence; do
+    for EXE in helm helmfile kind kubectl telepresence; do
         command -v "$EXE" >/dev/null 2>&1 || { echo "$EXE is required but cannot be found in PATH"; exit 1; }
     done
 }
@@ -20,7 +20,6 @@ stop_kind() {
     kind delete cluster --name "$CLUSTER"
 }
 
-
 check_prerequisites
 start_kind 
 
@@ -28,8 +27,5 @@ if [[ "$CLEANUP" == "true" ]]; then
     trap stop_kind EXIT
 fi
 
-telepresence connect --no-report -- ginkgo "$@"
-
-
-
+telepresence connect --no-report -- go test -v --tags="tests e2e" ./...
