--- e2e/tests/postgres/postgres_suite_test.go
@@ -1,20 +1,0 @@
-// Copyright 2021-2022 Zenauth Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-package postgres_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-
-	"github.com/cerbos/cerbos/e2e/common"
-)
-
-func TestPostgres(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Postgres Suite")
-}
-
-var _ = SynchronizedBeforeSuite(common.SBSDeploy, func(arg []byte) {})
